+++ app/src/renderer/vuex/getters.js
@@ -60,7 +60,6 @@ export const oldUnbondingAtoms = (state, getters) => {
 }
 export const committedDelegations = state => state.delegation.committedDelegates
 export const delegates = state => state.delegates
-export const shoppingCart = state => state.delegation.delegates
 export const validators = state => state.validators.validators
 export const keybase = state => state.keybase.identities
 export const pool = state => state.pool