+++ app/src/renderer/vuex/store.js
@@ -121,12 +121,5 @@ function loadPersistedState({ state, commit }, { password }) {
       }
     })
     this.replaceState(state)
-
-    // add all delegates the user has bond with already to the cart
-    state.delegates.delegates
-      .filter(d => state.delegation.committedDelegates[d.operator_address])
-      .forEach(d => {
-        commit(`addToCart`, d)
-      })
   }
 }